+++ university/src/main/java/ua/foxminded/university/service/GenericService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface GenericService <T> {
     void add(T t);
-    void deleteById(int id);
-    T getById(int id);
+    void deleteById(long id);
+    T getById(long id);
     List<T> getAll();
 }
